+++ .old/screens/equipment.rs
