+++ fallout-2d20-gui/src/main.rs
@@ -3,43 +3,77 @@ mod config;
 mod screens;
 mod character;
 mod theme;
-mod main2;
-mod screens2;
 
-use db::Db;
+use std::os::raw::c_void;
+use glow::HasContext;
 use sdl2::video::{ GLProfile,Window };
 use imgui_sdl2::ImguiSdl2;
 use imgui_opengl_renderer::Renderer;
-use imgui::{ Ui };
-use std::os::raw::c_void;
-use std::path::PathBuf;
+use imgui::Ui;
 use anyhow::Result;
-use glow::HasContext;
-use config::{ load_config, save_config, AppConfig };
-use theme::{ Theme, BAR_HEIGHT, THEME_CAPITAL, THEME_COMMONWEALTH, THEME_MOJAVE, THEME_NUCLEAR_SHADOW, THEME_NUCLEAR_WINTER, apply_theme};
-use screens::main_menu::render_main_menu;
-use screens::new_character::{ NewCharacterState, render_new_character, render_text_wrapped };
-use screens::special::{ render_special, SpecialState, MutantType };
-use screens::skills::{ render_skills, SkillsState, sync_trait_effects };
-use screens::perks::{ render_perks, PerksState, load_perks, render_perk_resolution, PerkResolutionPopup };
-use screens::stats::{ render_stats, ComputedStats, compute_stats };
-use screens::equipment::{ render_equipment, EquipmentState };
-use screens::review::render_review;
+
+use crate::{
+    character::{Character, Party, Player}, config::{AppConfig, db_path, load_config, save_config}, db::Db, screens::perk_select::PerkResolutionPopup, theme::{BAR_HEIGHT, THEMES, apply_theme, render_text_wrapped}
+};
+use crate::screens::main_menu::render_main_menu;
+use crate::screens::origin_select::{render_origin_select, OriginState};
+use crate::screens::special_assignment::{render_special_assignment, SpecialState};
+use crate::screens::skill_assignment::{render_skill_assignment, SkillState};
+use crate::screens::perk_select::{render_perk_select, PerkState, render_perk_resolution};
+use crate::screens::stat_calculation::render_stat_calculation;
+use crate::screens::background_select::{render_background_select, BackgroundState};
+use crate::screens::character_review::render_character_review;
 
 #[derive(Debug, Clone, PartialEq)]
-enum AppScreen {
+pub enum AppScreen {
     MainMenu,
-    NewCharacter,
     LoadCharacter,
     ImportCharacter,
-    Special,
-    Skills,
-    Perks,
-    Stats,
-    Equipment,
-    Review
+    OriginSelect,
+    SpecialAssignment,
+    SkillAssignment,
+    PerkSelect,
+    StatCalculation,
+    BackgroundSelect,
+    CharacterReview,
+    CharacterSheet,
 }
 
+pub const BUILD_SCREENS: &[(AppScreen, &str)] = &[
+    (AppScreen::OriginSelect, "Origin"),
+    (AppScreen::SpecialAssignment, "SPECIAL"),
+    (AppScreen::SkillAssignment, "Skills"),
+    (AppScreen::PerkSelect, "Perks"),
+    (AppScreen::StatCalculation, "Stats"),
+    (AppScreen::BackgroundSelect, "Background"),
+    (AppScreen::CharacterReview, "Review"),
+];
+
+const VERSION: &str = "0.1.9-alpha.1";
+const DATE: &str = "20260409";
+
+pub fn screen_unlocked(
+    screen: &AppScreen,
+    origin: &OriginState,
+    special: &SpecialState,
+    skill: &SkillState,
+    perk: &PerkState,
+    background: &BackgroundState,
+    character: &Character,
+) -> bool {
+    match screen {
+        AppScreen::OriginSelect => true,
+        AppScreen::SpecialAssignment => origin.is_complete(),
+        AppScreen::SkillAssignment => special.is_complete(character),
+        AppScreen::PerkSelect => skill.is_complete(character),
+        AppScreen::StatCalculation => perk.is_complete(),
+        AppScreen::BackgroundSelect => special.is_complete(character) && skill.is_complete(character) && perk.is_complete(),
+        AppScreen::CharacterReview => background.is_complete(),
+        _ => false,
+    }
+}
+
+//build a placeholder window
 fn render_placeholder(ui: &Ui, window: &Window, title: &str, screen: &mut AppScreen) {
     let (win_w, win_h) = window.size();
     let w = 500.0_f32;
@@ -66,14 +100,15 @@ fn render_placeholder(ui: &Ui, window: &Window, title: &str, screen: &mut AppScr
 }
 
 fn main() -> Result<()> {
+    //build the window
     let sdl_context = sdl2::init().map_err(|e| anyhow::anyhow!(e))?;
     let video_subsystem = sdl_context.video().map_err(|e| anyhow::anyhow!(e))?;
 
     video_subsystem.gl_attr().set_context_profile(GLProfile::Core);
-    video_subsystem.gl_attr().set_context_version(3, 2);
+    video_subsystem.gl_attr().set_context_version(3,2);
 
     let window = video_subsystem
-        .window("Fallout 2d20 Character Manager", 1900, 950)
+        .window(&format!("Fallout 2d20 Character Manager v{}",VERSION), 1900, 950)
         .position_centered()
         .opengl()
         .resizable()
@@ -88,14 +123,18 @@ fn main() -> Result<()> {
         })
     };
 
-    let themes: [&Theme; 5] = [&THEME_CAPITAL, &THEME_MOJAVE, &THEME_COMMONWEALTH, &THEME_NUCLEAR_WINTER, &THEME_NUCLEAR_SHADOW];
-
+    //load the user config
     let cfg = load_config();
-    let mut current_theme = cfg.theme_index.min(themes.len() - 1);
+    //load the theme from the user config
+    let mut current_theme = cfg.theme_index.min(THEMES.len() - 1);
 
+    //creates the context for imgui functions
     let mut imgui = imgui::Context::create();
-    apply_theme(&mut imgui, themes[current_theme]);
-    //imgui.set_ini_filename(None);
+    //applies the theme from the user config
+    apply_theme(&mut imgui, THEMES[current_theme]);
+    //sets the path for the db
+    let db_path = &db_path(cfg.db_path);
+    //sets the file path for imgui ini file
     let ini_path = std::env::current_exe()
         .unwrap()
         .parent()
@@ -103,41 +142,151 @@ fn main() -> Result<()> {
         .join("imgui.ini");
     imgui.set_ini_filename(ini_path);
 
+    //create the renderer
     let mut imgui_sdl2 = ImguiSdl2::new(&mut imgui, &window);
     let renderer = Renderer::new(&mut imgui, |s| {
         video_subsystem.gl_get_proc_address(s) as *const c_void
     });
 
+    //create the event pump
     let mut event_pump = sdl_context.event_pump().map_err(|e| anyhow::anyhow!(e))?;
 
+    //create the main menu
     let mut screen = AppScreen::MainMenu;
     let mut selected_menu_item: i32 = 0;
     let menu_items = ["New Character", "Load Character", "Import Character", "Quit"];
-    
+
+    //since we just set our theme, pending theme can be set to the same thing
+    //we'll check this in the loop every frame to determine if the theme needs
+    //  to be updated
     let mut pending_theme: Option<usize> = Some(current_theme);
 
-    let mut show_about = false;
-    let mut new_char_state: Option<NewCharacterState> = None;
-    let mut special_state: Option<SpecialState> = None;
-    let mut skills_state: Option<SkillsState> = None;
-    let mut perks_state: Option<PerksState> = None;
-    let mut perk_resolution: Option<PerkResolutionPopup> = None;
-    let mut equipment_state: Option<EquipmentState> = None;
-    let mut stats_state: Option<ComputedStats> = None;
-    
-    let db_path = &config::db_path(cfg.db_path);
-    //let db_path_str = &db_path.display().to_string();
-    /*
-    let db_path = std::env::current_exe()
-        .unwrap()
-        .parent()
-        .unwrap()
-        .join("fallout_2d20.db");
-    */
+    //create the db if it doesn't exist, which avoids errors
+    //if it's a new db, it's going to result in blank entries for origin select
     std::fs::create_dir_all(db_path.parent().unwrap())?;
     let db = Db::connect(&format!("sqlite:{}", db_path.display()))?;
 
+    //initializing all the states on first load as None
+    let mut show_about = false;
+    let mut player = Player::new();
+    //let mut party = Party::new();
+    let mut character = Character::new(player, None);
+    let mut party: Option<Party> = None;
+    let mut origin = OriginState::new(&db);
+    let mut special = SpecialState::new();
+    let mut skill = SkillState::new(character.clone());
+    let mut perk = PerkState::new(&db, &character);
+    let mut perk_resolution: Option<PerkResolutionPopup> = None;
+    let mut background = BackgroundState::new(&db);
+
+    //start the render loop
     'main: loop {
+        //function for handling the tabbed windows for the builder:
+        pub fn render_tab_bar(
+            ui: &Ui,
+            current: AppScreen,
+            screen: &mut AppScreen,
+            origin: &OriginState,
+            special: &SpecialState,
+            skill: &SkillState,
+            perk: &PerkState,
+            background: &BackgroundState,
+            character: &Character,
+        ) {
+            //evenly space the tabs
+            let tab_w = ui.content_region_avail()[0] / BUILD_SCREENS.len() as f32;
+
+            //establish which tab is the current tab and which are unlocked
+            for (target, label) in BUILD_SCREENS {
+                let is_current  = target == &current;
+                let is_unlocked = screen_unlocked(target, origin, special, skill, perk, background, &character);
+
+                //highlight the current tab
+                if is_current {
+                    let color = ui.push_style_color(
+                        imgui::StyleColor::Button,
+                        ui.style_color(imgui::StyleColor::ButtonActive),
+                    );
+                    ui.set_next_item_width(tab_w);
+                    ui.button(label);
+                    drop(color);
+                    //render the tab normally
+                } else if is_unlocked {
+                    ui.set_next_item_width(tab_w);
+                    if ui.button(label) {
+                        *screen = target.clone();
+                    }
+                } else {
+                    //disable the tab
+                    let color = ui.push_style_color(
+                        imgui::StyleColor::Text,
+                        ui.style_color(imgui::StyleColor::TextDisabled),
+                    );
+                    let color2 = ui.push_style_color(
+                        imgui::StyleColor::Button,
+                        ui.style_color(imgui::StyleColor::FrameBg),
+                    );
+                    ui.set_next_item_width(tab_w);
+                    ui.button(label);
+                    drop(color);
+                    drop(color2);
+                }
+
+                ui.same_line_with_spacing(0.0, 0.0);
+            }
+
+            ui.new_line();
+            ui.separator();
+            ui.spacing();
+        }
+
+        //function for rendering forward/back buttons
+        pub fn render_nav_footer(
+            ui: &Ui,
+            h: f32,
+            current: AppScreen,
+            screen: &mut AppScreen,
+            origin: &OriginState,
+            special: &SpecialState,
+            skill: &SkillState,
+            perk: &PerkState,
+            background: &BackgroundState,
+            character: &Character,
+        ) {
+            //figure out which tab/screen we're on
+            let idx = BUILD_SCREENS.iter().position(|(s, _)| s == &current).unwrap_or(0);
+
+            //define the previous and next screens
+            let prev = idx.checked_sub(1).map(|i| &BUILD_SCREENS[i].0);
+            let next = BUILD_SCREENS.get(idx + 1).map(|(s, _)| s);
+
+            ui.separator();
+            ui.spacing();
+            ui.set_cursor_pos([16.0, h - 36.0]);
+
+            //if there's a previous screen, create a back button and point to it
+            if let Some(prev_screen) = prev {
+                if ui.button("< Back") {
+                    *screen = prev_screen.clone();
+                }
+                ui.same_line();
+            }
+
+            if let Some(next_screen) = next {
+                let unlocked = screen_unlocked(next_screen, origin, special, skill, perk, background, &character);
+                //disable the next button if it's not unlocked
+                if !unlocked {
+                    let c = ui.push_style_color(imgui::StyleColor::Text, ui.style_color(imgui::StyleColor::TextDisabled));
+                    let c2 = ui.push_style_color(imgui::StyleColor::Button, ui.style_color(imgui::StyleColor::FrameBg));
+                    ui.button("Next >");
+                    drop(c); drop(c2);
+                } else if ui.button("Next >") {
+                    *screen = next_screen.clone();
+                }
+            }
+        }
+
+        //listen for events and handle them
         for event in event_pump.poll_iter() {
             imgui_sdl2.handle_event(&mut imgui, &event);
             if let sdl2::event::Event::Quit { .. } = event {
@@ -145,16 +294,19 @@ fn main() -> Result<()> {
             }
         }
         
+        //if pending theme is not None, apply it and make it None
         if let Some(t) = pending_theme.take() {
-            apply_theme(&mut imgui, themes[t]);
+            apply_theme(&mut imgui, THEMES[t]);
         }
 
+        //create the frame for rendering stuff
         imgui_sdl2.prepare_frame(imgui.io_mut(), &window, &event_pump.mouse_state());
         let ui = imgui.frame();
 
-        // ── Always-visible theme bar ──────────────────────────────────────────────────
+        //get the window size
         let (win_w, _win_h) = window.size();
 
+        //create the theme bar at the very top
         ui.window("##theme_bar")
             .title_bar(false)
             .resizable(false)
@@ -165,15 +317,19 @@ fn main() -> Result<()> {
             .position([0.0, 0.0], imgui::Condition::Always)
             .build(|| {
                 ui.set_cursor_pos([8.0, 7.0]);
-                ui.text_colored(themes[current_theme].text_dim, "Theme:");
+                ui.text_colored(THEMES[current_theme].text_dim, "Theme:");
                 ui.same_line();
-                for (i, theme) in themes.iter().enumerate() {
+                //iterating through the themes to draw radio buttons
+                for (i, theme) in THEMES.iter().enumerate() {
+                    //checks if the radio button is clicked for the theme
                     if ui.radio_button_bool(theme.name, current_theme == i) {
+                        //sets the theme, then writes it to the config
                         current_theme = i;
                         pending_theme = Some(i);
                         save_config(&AppConfig { theme_index: i, db_path: db_path.to_path_buf() });
                     }
-                    if i < themes.len() - 1 {
+                    if i < THEMES.len() - 1 {
+                        //doesn't move to the next line unless we're at the end of the themes
                         ui.same_line();
                     }
                 }
@@ -181,35 +337,37 @@ fn main() -> Result<()> {
                 let button_w = 60.0_f32;
                 let button_x = win_w as f32 - button_w - 8.0;
                 ui.set_cursor_pos([button_x, 4.0]);
+                //set the show_about flag if clicked
                 if ui.button("About") {
-                    show_about = true; // open or re-center
+                    show_about = true; 
                 }
             });
 
+        //render about window if the flag is set
         if show_about {
             let (win_w, win_h) = window.size();
             let aw = 400.0_f32;
             let ah = 220.0_f32;
             let center = [(win_w as f32 - aw) * 0.5, (win_h as f32 - ah) * 0.5];
 
-            // Only force position when first opened or re-centered (not every frame)
+            //center the about window when it's opened or the button is clicked again (not every frame)
             let condition = if ui.is_mouse_released(imgui::MouseButton::Left) {
                 imgui::Condition::Appearing
             } else {
                 imgui::Condition::Appearing
             };
 
+            //rendering the about window
             ui.window("##about")
                 .title_bar(false)
                 .resizable(false)
-                .movable(true)          // draggable
+                .movable(true)
                 .collapsible(false)
                 .size([aw, ah], imgui::Condition::Always)
-                .position(center, imgui::Condition::Once) // Once = only set pos on first appear
-                //.bring_current_window_to_display_front()  // always on top
+                .position(center, imgui::Condition::Once)
                 .bring_to_front_on_focus(true)
                 .build(|| {
-                    // Title row with X button
+                    //title with X
                     let close_x = aw - 28.0;
                     ui.text("About");
                     ui.same_line_with_pos(close_x);
@@ -221,9 +379,9 @@ fn main() -> Result<()> {
 
                     ui.text("fallout 2d20 character manager");
                     ui.spacing();
-                    render_text_wrapped(true, false, ui, "v0.1.9, 20260408", 16.0, aw - 32.0);
+                    render_text_wrapped(true, false, ui, &format!("v{}, {}", VERSION, DATE), 16.0, aw - 32.0);
                     ui.spacing();
-                    ui.text_wrapped("A character creation and management tool for the 2d20 ttrpg system.");
+                    ui.text_wrapped("a character creation and management tool for the 2d20 ttrpg system.");
                     ui.text_colored([0.90, 0.10, 0.50, 1.00], "by josh");
                     ui.spacing();
                     ui.separator();
@@ -232,199 +390,79 @@ fn main() -> Result<()> {
                 });
         }
 
-        // ── Screen content (offset below the bar) ────────────────────────────────────
+        render_tab_bar(ui, screen.clone(), &mut screen, &origin, &special, &skill, &perk, &background, &character);
+
         match screen {
-            AppScreen::MainMenu => {
+/*--------*/AppScreen::MainMenu => {
                 render_main_menu(&ui, &window, &mut screen, &mut selected_menu_item, &menu_items);
             }
-            AppScreen::NewCharacter => {
-                let state = new_char_state.get_or_insert_with(|| NewCharacterState::load(&db));
-                render_new_character(&ui, &window, state, &mut screen, &db);
-                if screen == AppScreen::MainMenu {
-                    new_char_state = None;
-                }
+/*--------*/AppScreen::OriginSelect => {
+                let state = &mut origin;
+                let h = render_origin_select(&ui, &window, state, &db, &mut character);
+                render_nav_footer(ui, h, screen.clone(), &mut screen, &origin, &special, &skill, &perk, &background, &character);
             }
-            AppScreen::LoadCharacter => {
-                render_placeholder(&ui, &window, "Load Character", &mut screen);
+/*--------*/AppScreen::SpecialAssignment => {
+                let state = &mut special.update(&character);
+                let h = render_special_assignment(&ui, &window, state, &db, &mut character);
+                render_nav_footer(ui, h, screen.clone(), &mut screen, &origin, &special, &skill, &perk, &background, &character);
             }
-            AppScreen::ImportCharacter => {
-                render_placeholder(&ui, &window, "Import Character", &mut screen);
+/*--------*/AppScreen::SkillAssignment => {
+                let state = &mut skill;
+                let h = render_skill_assignment(&ui, &window, state, &db, &mut character);
+                render_nav_footer(ui, h, screen.clone(), &mut screen, &origin, &special, &skill, &perk, &background, &character);
             }
-            // In the match:
-            AppScreen::Special => {
-                // Pull is_gifted and mutant_type from new_char_state
-                let (is_gifted, mutant_type) = new_char_state
-                    .as_ref()
-                    .map(|s| (s.has_gifted_trait(), s.mutant_type()))
-                    .unwrap_or((false, MutantType::None));
-
-                let state = special_state.get_or_insert_with(|| {
-                    SpecialState::new(is_gifted, mutant_type)
-                });
-
-                state.is_gifted = is_gifted;
-                state.mutant_type = mutant_type;
-
-                if !is_gifted {
-                    state.gifted_selected = [false; 7];
-                }
-
-                render_special(&ui, &window, state, &mut screen);
-                if screen == AppScreen::MainMenu {
-                    special_state = None;
-                }
-            }
-            AppScreen::Skills => {
-                // Sync intelligence from completed SPECIAL state
-                let intelligence = special_state
-                    .as_ref()
-                    .map(|s| s.display_value(crate::screens::special::I))
-                    .unwrap_or(5);
-                let level = new_char_state.as_ref().map(|s| s.level).unwrap_or(1);
-
-                let state = skills_state.get_or_insert_with(|| {
-                    SkillsState::new(intelligence, level)
-                });
-
-                state.intelligence = intelligence;
-                state.level = level;
-
-                // Sync trait effects each frame
-                if let Some(nc) = &new_char_state {
-                    let selected_ids: Vec<i32> = nc.traits.iter().enumerate()
-                        .filter(|(i, _)| nc.selected_traits.get(*i).copied().unwrap_or(false))
-                        .map(|(_, t)| t.id as i32)
-                        .collect();            
-                    sync_trait_effects(state, &selected_ids, nc.is_ghoul);
-                }
-
-                render_skills(&ui, &window, state, &mut screen);
-                if screen == AppScreen::MainMenu {
-                    skills_state = None;
-                }
-            }
-            AppScreen::Perks => {
-                let special_display = special_state
-                    .as_ref()
-                    .map(|s| std::array::from_fn(|i| s.display_value(i).into()))
-                    .unwrap_or([5; 7]);
-                let level = new_char_state.as_ref().map(|s| s.level).unwrap_or(1).into();
-                let is_ghoul = new_char_state.as_ref().map(|s| s.is_ghoul).unwrap_or(false);
-                let is_super_mutant = new_char_state.as_ref()
-                    .map(|s| s.mutant_type() != MutantType::None)
-                    .unwrap_or(false);
-                let perk_trait = new_char_state.as_ref()
-                    .map(|s| s.traits.iter().enumerate()
-                        .any(|(i, t)| t.id == 10 && s.selected_traits.get(i).copied().unwrap_or(false)))
-                    .unwrap_or(false);
-
-                let state = perks_state.get_or_insert_with(|| {
-                    let all_perks = load_perks(&db);
-                    PerksState::new(all_perks, level, special_display,
-                        is_ghoul, false, is_super_mutant, false, perk_trait)
-                });
-
-                // Sync mutable context each frame
-                state.level = level;
-                state.special = special_display;
-                state.is_ghoul = is_ghoul;
-                state.is_super_mutant = is_super_mutant;
-                state.perk_trait = perk_trait;
-
-                render_perks(&ui, &window, state, &mut screen, perk_resolution.is_some());
-                // Open resolution popup if a special perk was just taken
-                if let Some(pid) = state.pending_resolution.take() {
-                    let pname = state.all_perks.iter()
-                        .find(|p| p.id == pid)
-                        .map(|p| p.name.as_str())
-                        .unwrap_or("");
-                    if let Some(popup) = state.begin_resolve(pid, pname) {
+/*--------*/AppScreen::PerkSelect => {
+                let state = &mut perk;
+                let h = render_perk_select(&ui, &window, state, &mut screen, &db, &mut character, perk_resolution.is_some());
+                //resolution popup
+                if let Some((p_id, add)) = state.pending_resolution.take() {
+                    let perk = state.perks.iter().find(|p| p.id == p_id).unwrap();
+                    if let Some(popup) = state.begin_resolve(perk, add) {
                         perk_resolution = Some(popup);
                     }
                 }
-                // Render resolution popup if open
+                //render popup
                 if let Some(popup) = &mut perk_resolution {
-                    let special_max: [i32; 7] = std::array::from_fn(|i| {
-                        special_state.as_ref().map(|s| s.stat_max(i)).unwrap_or(10)
-                    });
-                    let result = render_perk_resolution(
-                        &ui, &window, popup,
-                        &mut state.special,
-                        &special_max,
-                        skills_state.as_mut().unwrap(),
-                        special_state.as_mut().unwrap(),
-                    );
+                    let result = render_perk_resolution(ui, &window, popup, state, &mut character);
                     match result {
                         Some(false) => {
-                            // Cancelled — remove the perk
-                            let pid = popup.perk_id;
-                            state.remove_perk(pid);
+                            let perk_index = character.clone().perks.iter().position(|p| p.id == popup.perk_id).unwrap();
+                            character.perks.remove(perk_index);
                             perk_resolution = None;
                         }
                         Some(true) => {
                             perk_resolution = None;
                         }
-                        None => {} // still open
+                        None => {} //it's still open so don't do anything
                     }
                 }
-                if screen == AppScreen::MainMenu {
-                    perks_state = None;
-                }
+                render_nav_footer(ui, h, screen.clone(), &mut screen, &origin, &special, &skill, &perk, &background, &character);
             }
-            AppScreen::Stats => {
-                let special = special_state.as_ref().unwrap();
-                let traits = new_char_state.as_ref().unwrap();
-                let perks = perks_state.as_ref().unwrap();
-                let state = stats_state.get_or_insert_with(|| {
-                    compute_stats(special, traits, perks)
-                });
-                render_stats(
-                    &ui, &window,
-                    special,
-                    skills_state.as_ref().unwrap(),
-                    perks,
-                    traits,
-                    &mut screen,
-                    state,
-                );
-                if screen == AppScreen::MainMenu {
-                    stats_state = None;
-                }
+/*--------*/AppScreen::StatCalculation => {
+                let h = render_stat_calculation(&ui, &window, &special, &skill, &mut character);
+                render_nav_footer(ui, h, screen.clone(), &mut screen, &origin, &special, &skill, &perk, &background, &character);
             }
-            AppScreen::Equipment => {
-                let origin_id = new_char_state.as_ref()
-                    .and_then(|s| s.selected_origin_id());
-                let state = equipment_state.get_or_insert_with(|| {
-                    let all_backgrounds = screens::equipment::load_backgrounds(&db);
-                    EquipmentState::new(all_backgrounds)
-                });
-
-                if state.origin_id != origin_id {
-                    state.origin_id = origin_id;
-                    state.reset_selection();
-                }
-
-                render_equipment(&ui, &window, state, &db, &mut screen);
-
-                if screen == AppScreen::MainMenu {
-                    equipment_state = None;
-                }
+/*--------*/AppScreen::BackgroundSelect => {
+                let state = &mut background;
+                let h = render_background_select(&ui, &window, state, &db, &mut character);
+                render_nav_footer(ui, h, screen.clone(), &mut screen, &origin, &special, &skill, &perk, &background, &character);
             }
-            AppScreen::Review => {
-                render_review(
-                    &ui, &window,
-                    new_char_state.as_ref().unwrap(),
-                    special_state.as_ref().unwrap(),
-                    skills_state.as_ref().unwrap(),
-                    perks_state.as_ref().unwrap(),
-                    stats_state.as_ref().unwrap(),
-                    equipment_state.as_ref().unwrap(),
-                    &mut screen,
-                    themes[current_theme],
-                    &db,
-                );
+/*--------*/AppScreen::CharacterReview => {
+                let h = render_character_review(&ui, &window, &db, &character);
+                render_nav_footer(ui, h, screen.clone(), &mut screen, &origin, &special, &skill, &perk, &background, &character);
             }
-
+/*--------*/AppScreen::CharacterSheet => {
+                render_placeholder(&ui, &window, "sheet", &mut screen);
+                //let state = &mut special;
+                //let h = render_special_assignment(&ui, &window, state, &mut screen, &db, &mut character);
+                //render_nav_footer(ui, h, screen.clone(), &mut screen, &origin, special, skill, perk, background, &character);
+            }
+/*--------*/AppScreen::LoadCharacter => {
+                render_placeholder(&ui, &window, "load", &mut screen);
+            }
+/*--------*/AppScreen::ImportCharacter => {
+                render_placeholder(&ui, &window, "import", &mut screen);
+            }
         }
 
         unsafe {
@@ -437,6 +475,5 @@ fn main() -> Result<()> {
 
         window.gl_swap_window();
     }
-
     Ok(())
 }
