+++ fallout-2d20-gui/src/screens/main_menu.rs
@@ -1,7 +1,6 @@
 use imgui::Ui;
 use sdl2::video::Window;
 use crate::AppScreen;
-use crate::screens::new_character::render_text_wrapped;
 
 pub fn render_main_menu(
     ui: &Ui,
@@ -10,6 +9,7 @@ pub fn render_main_menu(
     selected: &mut i32,
     items: &[&str],
 ) {
+    //get window dimmensions
     let (win_w, win_h) = window.size();
     let menu_w = 340.0_f32;
     let menu_h = 320.0_f32;
@@ -25,9 +25,10 @@ pub fn render_main_menu(
             imgui::Condition::Always,
         )
         .build(|| {
-            // Title
-            let title = "FALLOUT 2d20 COMPANION";
+            //title
+            let title = "fallout 2d20 companion";
             let title_w = ui.calc_text_size(title)[0];
+            //center the text on the top of the menu
             ui.set_cursor_pos([(menu_w - title_w) * 0.5, 24.0]);
             ui.text(title);
 
@@ -35,7 +36,7 @@ pub fn render_main_menu(
             ui.spacing();
             ui.spacing();
 
-            // Keyboard nav: arrow keys move selection
+            //navigate the menu with the arrows and enter/space
             if ui.is_window_focused() {
                 if ui.is_key_pressed_no_repeat(imgui::Key::DownArrow) {
                     *selected = (*selected + 1).min(items.len() as i32 - 1);
@@ -43,20 +44,21 @@ pub fn render_main_menu(
                 if ui.is_key_pressed_no_repeat(imgui::Key::UpArrow) {
                     *selected = (*selected - 1).max(0);
                 }
-                if ui.is_key_pressed_no_repeat(imgui::Key::Enter) || ui.is_key_pressed(imgui::Key::Space) {
+                if ui.is_key_pressed_no_repeat(imgui::Key::Enter) || ui.is_key_pressed_no_repeat(imgui::Key::Space) {
                     handle_selection(*selected, screen);
                 }
             }
 
-            // Render each item as a selectable
             for (i, &label) in items.iter().enumerate() {
                 let is_selected = *selected == i as i32;
+                //add an arrow to the label of the active button
                 let display = if is_selected {
                     format!("  > {}  ", label)
                 } else {
                     format!("    {}  ", label)
                 };
 
+                //align the buttons centered
                 let item_w = menu_w - 40.0;
                 let cursor_x = (menu_w - item_w) * 0.5;
                 let y = ui.cursor_pos()[1];
@@ -65,17 +67,15 @@ pub fn render_main_menu(
                 if ui.selectable_config(&display)
                     .selected(is_selected)
                     .size([item_w, 36.0])
-                    .build()
-                {
+                    .build() {
                     *selected = i as i32;
                     handle_selection(i as i32, screen);
                 }
 
-                // Hover sets selection too
+                //also select the button on hover (so that space/enter/click navigates)
                 if ui.is_item_hovered() {
                     *selected = i as i32;
                 }
-
                 ui.spacing();
             }
 
@@ -83,18 +83,16 @@ pub fn render_main_menu(
             ui.spacing();
             ui.separator();
 
-            // Footer hint
-            let hint = "arrow keys / click to select  |  enter to confirm";
-            //let hint_w = ui.calc_text_size(hint)[0];
-            //ui.set_cursor_pos([(menu_w - hint_w) * 0.5, menu_h - 28.0]);
-            render_text_wrapped(true, false, ui, hint, 16.0, menu_w - 16.0);
-            //ui.text_disabled(hint);
+            let hint = "arrow/hover select | enter/space confirm";
+            let hint_w = ui.calc_text_size(hint)[0];
+            ui.set_cursor_pos([(menu_w - hint_w) * 0.5, menu_h - 28.0]);
+            ui.text_disabled(hint);
         });
 }
 
 fn handle_selection(selected: i32, screen: &mut AppScreen) {
     match selected {
-        0 => *screen = AppScreen::NewCharacter,
+        0 => *screen = AppScreen::OriginSelect,
         1 => *screen = AppScreen::LoadCharacter,
         2 => *screen = AppScreen::ImportCharacter,
         3 => std::process::exit(0),
