+++ fallout-2d20-gui/src/screens/mod.rs
@@ -1,8 +1,9 @@
 pub mod main_menu;
-pub mod new_character;
-pub mod special;
-pub mod skills;
-pub mod perks;
-pub mod stats;
-pub mod equipment;
-pub mod review;
+pub mod origin_select;
+pub mod special_assignment;
+pub mod skill_assignment;
+pub mod perk_select;
+pub mod stat_calculation;
+pub mod background_select;
+pub mod character_review;
+pub mod character_sheet;
