+++ fallout-2d20-gui/src/screens/skill_assignment.rs
