+++ fallout-2d20-gui/src/screens/special_assignment.rs
