--- fallout-2d20-gui/src/screens2/mod.rs
@@ -1,9 +1,0 @@
-pub mod main_menu;
-pub mod origin_select;
-pub mod special_assignment;
-pub mod skill_assignment;
-pub mod perk_select;
-pub mod stat_calculation;
-pub mod background_select;
-pub mod character_review;
-pub mod character_sheet;
